pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "okx-news-scraper"
version = "0.1.0"
description = "CLI scraper for OKX announcements"
authors = [{ name = "Aleksandr Tarletskiy", email = "alexander.tarletsky@gmail.com" }]
dependencies = [
  "requests>=2.20",
  "beautifulsoup4>=4.6",
  "python-dateutil>=2.8",
]

# If you still want a console script:
[project.scripts]
okxnews = "okx_news_scraper.cli:main"